@@ -14,49 +14,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~14.2.3",
-    "@angular/cdk": "^14.1.3",
-    "@angular/common": "~14.2.3",
-    "@angular/compiler": "~14.2.3",
-    "@angular/core": "~14.2.3",
-    "@angular/forms": "~14.2.3",
-    "@angular/material": "^14.1.3",
-    "@angular/platform-browser": "~14.2.3",
-    "@angular/platform-browser-dynamic": "~14.2.3",
-    "@angular/router": "~14.2.3",
-    "@ng-idle/core": "^11.1.0",
-    "@ng-idle/keepalive": "^11.0.3",
-    "@progress/kendo-angular-common": "^13.5.0",
-    "@progress/kendo-angular-icons": "^13.5.0",
-    "@progress/kendo-angular-l10n": "^13.5.0",
-    "@progress/kendo-angular-layout": "^13.5.0",
-    "@progress/kendo-angular-progressbar": "^13.5.0",
+    "@angular/animations": "~18.1.2",
+    "@angular/cdk": "^18.1.2",
+    "@angular/common": "~18.1.2",
+    "@angular/compiler": "~18.1.2",
+    "@angular/core": "~18.1.2",
+    "@angular/forms": "~18.1.2",
+    "@angular/material": "^18.1.2",
+    "@angular/platform-browser": "~18.1.2",
+    "@angular/platform-browser-dynamic": "~18.1.2",
+    "@angular/router": "~18.1.2",
+    "@ng-idle/core": "^14.0.0",
+    "@ng-idle/keepalive": "^14.0.0",
+    "@progress/kendo-angular-common": "^16.5.0",
+    "@progress/kendo-angular-icons": "^16.5.0",
+    "@progress/kendo-angular-l10n": "^16.5.0",
+    "@progress/kendo-angular-layout": "^16.5.0",
+    "@progress/kendo-angular-progressbar": "^16.5.0",
     "@progress/kendo-licensing": "^1.3.2",
-    "angular-oauth2-oidc": "^12.1.0",
-    "angular2-text-mask": "^9.0.0",
+    "angular-imask": "^7.6.1",
+    "angular-oauth2-oidc": "^17.0.2",
     "bootstrap": "^5.1.3",
-    "eslint-plugin-deprecation": "~1.3.2",
+    "eslint-plugin-deprecation": "~3.0.0",
     "moment": "^2.29.4",
-    "ng-recaptcha": "^10.0.0",
-    "ngx-markdown": "^14.0.1",
+    "ng-recaptcha-2": "^14.0.0",
+    "ngx-markdown": "^18.0.0",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.2.3",
-    "@angular-eslint/builder": "~14.1.2",
-    "@angular-eslint/eslint-plugin": "~14.1.2",
-    "@angular-eslint/eslint-plugin-template": "~14.1.2",
-    "@angular-eslint/schematics": "~14.1.2",
-    "@angular-eslint/template-parser": "~14.1.2",
-    "@angular/cli": "~14.2.3",
-    "@angular/compiler-cli": "~14.2.3",
+    "@angular-devkit/build-angular": "~18.1.2",
+    "@angular-devkit/core": "^18.1.2",
+    "@angular-devkit/schematics": "^18.1.2",
+    "@angular-eslint/builder": "~18.1.0",
+    "@angular-eslint/eslint-plugin": "~18.1.0",
+    "@angular-eslint/eslint-plugin-template": "~18.1.0",
+    "@angular-eslint/schematics": "~18.1.0",
+    "@angular-eslint/template-parser": "~18.1.0",
+    "@angular/cli": "^18.1.2",
+    "@angular/compiler-cli": "~18.1.2",
+    "@progress/kendo-angular-buttons": "^16.5.0",
+    "@progress/kendo-angular-intl": "^16.5.0",
+    "@progress/kendo-svg-icons": "^3.0.0",
     "@types/jasmine": "~3.10.2",
     "@types/node": "^12.20.5",
-    "@typescript-eslint/eslint-plugin": "~5.36.2",
-    "@typescript-eslint/parser": "~5.36.2",
-    "eslint": "^8.23.1",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "@typescript-eslint/utils": "^7.17.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "~8.5.0",
     "eslint-plugin-import": "~2.26.0",
     "eslint-plugin-jsdoc": "~39.3.0",
@@ -71,6 +77,6 @@
     "ng-openapi-gen": "^0.21.2",
     "prettier": "^2.6.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.7.4"
+    "typescript": "5.4.5"
   }
 }
