@@ -4,9 +4,9 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import {MatNativeDateModule} from '@angular/material/core';
+import { MatNativeDateModule } from '@angular/material/core';
 import { MatDatepickerModule } from '@angular/material/datepicker'
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { CoreModule } from './core/core.module';
 import { ApiModule } from './core/api/api.module';
 import { OAuthModule } from 'angular-oauth2-oidc';
@@ -26,53 +26,52 @@ import { MAT_SELECT_SCROLL_STRATEGY_PROVIDER } from '@angular/material/select';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 
 @NgModule({
-  declarations: [AppComponent, OutageBannerComponent, OutageDialogComponent],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    MatDatepickerModule,
-    AppRoutingModule,
-    MatNativeDateModule,
-    BrowserAnimationsModule,
-    ReactiveFormsModule,
-    CoreModule,
-    ApiModule.forRoot({ rootUrl: '.' }),
-    NgIdleKeepaliveModule.forRoot(),
-    OAuthModule.forRoot({
-      resourceServer: {
-        sendAccessToken: true,
-        customUrlValidation: (url) =>
-          url.toLowerCase().includes('/api/') &&
-          !url.toLowerCase().endsWith('/configuration')
-      }
-    }),
-    LayoutModule,
-    ButtonsModule,
-    MatIconModule,
-    MatAutocompleteModule
-  ],
-  exports: [
-    MatIconModule
-  ],
-  providers: [
-    {
-      provide: APP_BASE_HREF,
-      useFactory: (s: PlatformLocation) => {
-        let result = s.getBaseHrefFromDOM();
-        if (result[result.length - 1] === '/') {
-          result = result.substring(0, result.length - 1);
-        }
-        return result;
-      },
-      deps: [PlatformLocation]
-    },
-    {
-      provide: ICON_SETTINGS,
-      useValue: { type: "font" },
-    },
-    ScriptService
-  ],
-  bootstrap: [AppComponent]
+    declarations: [AppComponent, OutageBannerComponent, OutageDialogComponent],
+    exports: [
+        MatIconModule
+    ],
+    bootstrap: [AppComponent],
+    imports: [
+        BrowserModule,
+        MatDatepickerModule,
+        AppRoutingModule,
+        MatNativeDateModule,
+        BrowserAnimationsModule,
+        ReactiveFormsModule,
+        CoreModule,
+        ApiModule.forRoot({ rootUrl: '.' }),
+        NgIdleKeepaliveModule.forRoot(),
+        OAuthModule.forRoot({
+            resourceServer: {
+                sendAccessToken: true,
+                customUrlValidation: (url) => url.toLowerCase().includes('/api/') &&
+                    !url.toLowerCase().endsWith('/configuration')
+            }
+        }),
+        LayoutModule,
+        ButtonsModule,
+        MatIconModule,
+        MatAutocompleteModule
+    ],
+    providers: [
+        {
+            provide: APP_BASE_HREF,
+            useFactory: (s: PlatformLocation) => {
+                let result = s.getBaseHrefFromDOM();
+                if (result[result.length - 1] === '/') {
+                    result = result.substring(0, result.length - 1);
+                }
+                return result;
+            },
+            deps: [PlatformLocation]
+        },
+        {
+            provide: ICON_SETTINGS,
+            useValue: { type: "font" },
+        },
+        ScriptService,
+        provideHttpClient(withInterceptorsFromDi())
+    ]
 })
 export class AppModule { }
 export class MaterialModule { }
