@@ -1,10 +1,14 @@
 import { Component, OnChanges, OnInit, SimpleChanges } from '@angular/core';
-import { Router } from '@angular/router';
+import { ActivatedRoute, NavigationEnd, NavigationStart, Router } from '@angular/router';
 import { FormCreationService } from '../../services/formCreation.service';
 import { CacheService } from '../../services/cache.service';
 import { ContactService } from 'src/app/core/api/services';
 import { LoginService } from '../../../core/services/login.service';
 import { first, map, mergeMap } from 'rxjs';
+import { DFAApplicationMainDataService } from '../../../feature-components/dfa-application-main/dfa-application-main-data.service';
+import { DFAClaimMainDataService } from '../../../feature-components/dfa-claim-main/dfa-claim-main-data.service';
+import { DFAProjectMainDataService } from '../../../feature-components/dfa-project-main/dfa-project-main-data.service';
+import { forEachChild } from 'typescript';
 
 @Component({
   selector: 'app-header',
@@ -13,9 +17,111 @@ import { first, map, mergeMap } from 'rxjs';
 })
 export class HeaderComponent implements OnInit {
   showLoginMatMenu: boolean;
+  showBreadcrumb: boolean = true;
 
   firstName: string = "";
   lastName: string = "";
+  appId = null;
+  projId = null;
+  claimId = null;
+
+  appviewedit = null;
+  projviewedit = null;
+  claimviewedit = null;
+
+  menuHeader: any = [];
+  appitems = [
+    {
+      label: 'Login',
+      link: '',
+      icon: 'pan_tool',
+      path: '/registration-method',
+      level: 0,
+      excluded: true,
+      islink: false
+    },
+    {
+      label: 'Create Application',
+      link: '',
+      icon: 'pan_tool',
+      path: 'dfa-application-main',
+      level: 2,
+      excluded: true,
+      islink: true
+    },
+    {
+      label: 'View Application',
+      link: 'dfa-application-main/appid',
+      icon: 'pan_tool',
+      path: 'dfa-application-main/',
+      level: 2,
+      excluded: false,
+      islink: true
+    },
+    {
+      label: 'Project Dashboard',
+      link: 'dfa-application/appid/projects',
+      icon: 'pan_tool',
+      path: '/projects',
+      level: 3,
+      excluded: false,
+      islink: true
+    },
+    {
+      label: 'Create Project',
+      link: 'dfa-project-main/prjid',
+      icon: 'pan_tool',
+      path: '/dfa-project-main/',
+      level: 4,
+      excluded: true,
+      islink: true
+    },
+    {
+      label: 'View Project',
+      link: 'dfa-project-view/prjid',
+      icon: 'pan_tool',
+      path: '/dfa-project-view/',
+      level: 4,
+      excluded: false,
+      islink: true
+    },
+    {
+      label: 'View Amendment',
+      link: 'dfa-project-amendment/prjid',
+      icon: 'pan_tool',
+      path: '/dfa-project-amendment/',
+      level: 4,
+      excluded: true,
+      islink: true
+    },
+    {
+      label: 'Claim Dashboard',
+      link: 'dfa-project/prjid/claims',
+      icon: 'pan_tool',
+      path: '/claims',
+      level: 5,
+      excluded: false,
+      islink: true
+    },
+    {
+      label: 'Claim Create',
+      link: 'dfa-claim-main/clid',
+      icon: 'pan_tool',
+      path: '/dfa-claim-main/',
+      level: 6,
+      excluded: true,
+      islink: true
+    },
+    {
+      label: 'View Claim',
+      link: 'dfa-claim-main/clid',
+      icon: 'pan_tool',
+      path: '/dfa-claim-main/',
+      level: 6,
+      excluded: true,
+      islink: true
+    },
+  ];
 
   constructor(
     public formCreationService: FormCreationService,
@@ -23,12 +129,81 @@ export class HeaderComponent implements OnInit {
     private cacheService: CacheService,
     private contactService: ContactService,
     private loginService: LoginService,
-  ) {}
+    private dfaApplicationMainDataService: DFAApplicationMainDataService,
+    private dfaProjectMainDataService: DFAProjectMainDataService,
+    private dfaClaimMainDataService: DFAClaimMainDataService,
+    private route: ActivatedRoute
+  ) {
+    router.events.subscribe(event => {
+      
+      if (event instanceof NavigationStart) {
+        var currentItem = null;
+
+        this.appitems.forEach(obj => {
+          if (event.url.indexOf(obj.path) > -1) {
+            currentItem = obj;
+            //item.link = 
+          }
+        });
+        
+        if (currentItem) {
+          this.showBreadcrumb = true;
+          this.menuHeader = [];
+          this.appitems.filter(m => m.level < currentItem.level && m.excluded == false).forEach(obj => {
+            this.menuHeader.push(obj);
+          });
+
+          const curItem = JSON.parse(JSON.stringify(currentItem)); ;
+          curItem.islink = false;
+          this.menuHeader.push(curItem);
+        }
+        else {
+          this.menuHeader = [];
+          this.showBreadcrumb = false;
+        }
+      }
+
+      if (event instanceof NavigationEnd) {
+        if (event.url.indexOf('/dfa-dashboard') > -1) {
+          this.menuHeader = [];
+          this.showBreadcrumb = true;
+        }
+      }
+
+    });
+  }
 
   ngOnInit(): void {
 
     this.showLoginMatMenu = false;
+    this.appId = this.dfaApplicationMainDataService.getApplicationId();
+    this.projId = this.dfaProjectMainDataService.getProjectId();
+    this.claimId = this.dfaClaimMainDataService.getClaimId();
 
+    this.dfaApplicationMainDataService.changeAppId.subscribe((appId) => {
+      this.appId = appId;
+    });
+
+    this.dfaProjectMainDataService.changeProjectId.subscribe((projId) => {
+      this.projId = projId;
+    })
+
+    this.dfaClaimMainDataService.changeClaimId.subscribe((claimId) => {
+      this.claimId = claimId;
+    });
+
+    this.dfaApplicationMainDataService.changeViewOrEdit.subscribe((appviewedit) => {
+      this.appviewedit = appviewedit == 'add' || appviewedit == 'edit' || appviewedit == 'update' ? true : false;
+    });
+
+    this.dfaProjectMainDataService.changeViewOrEdit.subscribe((projviewedit) => {
+      this.projviewedit = projviewedit == 'add' || projviewedit == 'edit' || projviewedit == 'update' ? true : false;
+    })
+
+    this.dfaClaimMainDataService.changeViewOrEdit.subscribe((claimviewedit) => {
+      this.claimviewedit = claimviewedit == 'add' || claimviewedit == 'edit' || claimviewedit == 'update' ? true : false;
+    });
+    
     // console.debug('[DFA] header.component about to call middle-tier API.');
 
     // 2024-08-12 EMCRI-487 waynezen; use new ContactService to get Business Name from Keycloak access token
@@ -46,6 +221,45 @@ export class HeaderComponent implements OnInit {
             })
         }
       });
+  }
+  
+  breadCrumbMain() {
+    this.menuHeader = [];
+    this.router.navigate(['dfa-dashboard']);
+  }
+
+  appitemsTravel;
+
+  breadCrumb(menuItem, c) {
+    var path = '';
+
+    if (menuItem) {
+      if (menuItem.link.indexOf('appid') > -1) {
+        path = menuItem.link.replace("appid", this.appId);
+      }
+
+      if (menuItem.link.indexOf('prjid') > -1) {
+        path = menuItem.link.replace("prjid", this.projId);
+      }
+
+      if (menuItem.link.indexOf('clid') > -1) {
+        path = menuItem.link.replace("clid", this.claimId);
+      }
+    }
+
+    this.router.navigate([path]);
+  }
+
+  menuChange(menuChange) {
+
+    if (menuChange.items) {
+
+      this.appitemsTravel = menuChange.items;
+      this.menuHeader.push({ label: menuChange.label, icon: 'keyboard_arrow_right', items: menuChange.items });
+      // this.menuHeader.push(menuChange);
+
+      //console.log('hasMultiMenuLabel');
+    }
   }
 
   homeButton(): void { }
