@@ -168,7 +168,7 @@ export class DFAProjectComponent
     lstProjects.forEach(x => {
       if (
         (x.status.toLowerCase() === "decision made"
-          || x.status.toLowerCase() === "closed: inactive" || x.status.toLowerCase() === "closed: withdrawn")
+          || x.status.toLowerCase() === "closed" || x.status.toLowerCase() === "closed: withdrawn")
         &&
         (x.dateFileClosed && (this.OneDayAgo >= new Date(x.dateFileClosed).getTime()))
       ) {
