@@ -1,17 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
 import { ReviewComponent } from './review.component';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { CustomPipeModule } from '../../core/pipe/customPipe.module';
 import { CoreModule } from '../../core/core.module';
-import { RecaptchaFormsModule, RecaptchaModule } from 'ng-recaptcha';
 import { CaptchaV2Component } from 'src/app/core/components/captcha-v2/captcha-v2.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatTableModule } from '@angular/material/table';
 import { DashReviewRoutingModule } from './review-routing.module';
+import { RecaptchaFormsModule, RecaptchaModule } from 'ng-recaptcha-2';
 
 @NgModule({
   declarations: [ReviewComponent, CaptchaV2Component],
