@@ -7,7 +7,7 @@ import {
 } from '@angular/core/testing';
 import { VerifiedRegistrationComponent } from './verified-registration.component';
 import { UntypedFormBuilder, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { TimeoutService } from 'src/app/core/services/timeout.service';
 import { MockTimeoutService } from 'src/app/unit-tests/mockTimeout.service';
@@ -15,6 +15,7 @@ import { MockExpiry } from 'src/app/unit-tests/mockExpiry.service';
 import { NgIdleKeepaliveModule } from '@ng-idle/keepalive';
 import { MatDialogModule } from '@angular/material/dialog';
 import { OAuthModule } from 'angular-oauth2-oidc';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('VerifiedRegistrationComponent', () => {
   let app: VerifiedRegistrationComponent;
@@ -24,24 +25,23 @@ describe('VerifiedRegistrationComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [VerifiedRegistrationComponent],
-      imports: [
-        ReactiveFormsModule,
-        HttpClientTestingModule,
+    declarations: [VerifiedRegistrationComponent],
+    imports: [ReactiveFormsModule,
         RouterTestingModule,
         NgIdleKeepaliveModule.forRoot(),
         MatDialogModule,
-        OAuthModule.forRoot()
-      ],
-      providers: [
+        OAuthModule.forRoot()],
+    providers: [
         UntypedFormBuilder,
         VerifiedRegistrationComponent,
         {
-          provide: TimeoutService,
-          useClass: MockTimeoutService
-        }
-      ]
-    }).compileComponents();
+            provide: TimeoutService,
+            useClass: MockTimeoutService
+        },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+}).compileComponents();
   }));
 
   beforeEach(() => {
