@@ -3,9 +3,10 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { LoginPageComponent } from './login-page.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { UntypedFormBuilder, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { LoginService } from '../core/services/login.service';
 import { OAuthModule } from 'angular-oauth2-oidc';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('LoginPageComponent', () => {
   let component: LoginPageComponent;
@@ -14,18 +15,17 @@ describe('LoginPageComponent', () => {
   beforeEach(waitForAsync(() => {
     let loginService: jasmine.SpyObj<LoginService>;
     TestBed.configureTestingModule({
-      declarations: [LoginPageComponent],
-      imports: [
-        RouterTestingModule,
+    declarations: [LoginPageComponent],
+    imports: [RouterTestingModule,
         ReactiveFormsModule,
-        HttpClientTestingModule,
-        OAuthModule.forRoot()
-      ],
-      providers: [
+        OAuthModule.forRoot()],
+    providers: [
         UntypedFormBuilder,
-        { provides: LoginService, useValue: loginService }
-      ]
-    }).compileComponents();
+        { provides: LoginService, useValue: loginService },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+}).compileComponents();
   }));
 
   beforeEach(() => {
