@@ -1,8 +1,9 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CanAddressComponent } from './can-address.component';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('CanAddressComponent', () => {
   let component: CanAddressComponent;
@@ -10,9 +11,10 @@ describe('CanAddressComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule, MatAutocompleteModule],
-      declarations: [CanAddressComponent]
-    }).compileComponents();
+    declarations: [CanAddressComponent],
+    imports: [MatAutocompleteModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}).compileComponents();
   }));
 
   beforeEach(() => {
