@@ -17,7 +17,7 @@ import { FormCreationService } from 'src/app/core/services/formCreation.service'
 import { BehaviorSubject, Observable, Subscription, catchError, mapTo, throwError } from 'rxjs';
 import { DirectivesModule } from '../../../../core/directives/directives.module';
 import { CustomValidationService } from 'src/app/core/services/customValidation.service';
-import { ApplicantOption, FileCategory, FileUpload, RequiredDocumentType, SmallBusinessOption, FarmOption } from 'src/app/core/api/models';
+import { ApplicantOption, FileCategory, FileUpload, RequiredDocumentType, SmallBusinessOption, FarmOption, InsuranceOption } from 'src/app/core/api/models';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
@@ -33,6 +33,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { MatDialog } from '@angular/material/dialog';
 import { DFAFileDeleteDialogComponent } from 'src/app/core/components/dialog-components/dfa-file-delete-dialog/dfa-file-delete.component';
 import { FileUploadWarningDialogComponent } from 'src/app/core/components/dialog-components/file-upload-warning-dialog/file-upload-warning-dialog.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-supporting-documents',
@@ -80,7 +81,9 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
   AppOptions = ApplicantOption;
   SmallBusinessOptions = SmallBusinessOption;
   FarmOptions = FarmOption;
+  InsuranceOptions = InsuranceOption;
   vieworedit: string = "";
+  isNoInsurance: boolean = false;
 
   constructor(
     @Inject('formBuilder') formBuilder: UntypedFormBuilder,
@@ -92,6 +95,7 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
     private http: HttpClient,
     private cd: ChangeDetectorRef,
     private dialog: MatDialog,
+    public _snackBar: MatSnackBar,
   ) {
     this.formBuilder = formBuilder;
     this.formCreationService = formCreationService;
@@ -104,6 +108,7 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
 
     this.dfaApplicationMainDataService.getDfaApplicationStart().subscribe(application => {
       if (application) {
+        this.isNoInsurance = (application.appTypeInsurance.insuranceOption == Object.keys(this.InsuranceOptions)[Object.values(this.InsuranceOptions).indexOf(this.InsuranceOptions.No)]);
         this.isResidentialTenant = (application.appTypeInsurance.applicantOption == Object.keys(this.AppOptions)[Object.values(this.AppOptions).indexOf(this.AppOptions.ResidentialTenant)]);
         this.isHomeowner = (application.appTypeInsurance.applicantOption == Object.keys(this.AppOptions)[Object.values(this.AppOptions).indexOf(this.AppOptions.Homeowner)]);
         this.isSmallBusinessOwner = (application.appTypeInsurance.applicantOption == Object.keys(this.AppOptions)[Object.values(this.AppOptions).indexOf(this.AppOptions.SmallBusinessOwner)]);
@@ -177,9 +182,13 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
     let supportingFiles = form.get('fileUploads')?.getRawValue();
     let applicantType = form.get('applicantType').value;
     const error={};
-    if (!supportingFiles || supportingFiles?.filter(x => x.requiredDocumentType === "InsuranceTemplate" && x.deleteFlag == false).length <= 0) {
+    if ((!supportingFiles || supportingFiles?.filter(x => x.requiredDocumentType === "InsuranceTemplate" && x.deleteFlag == false).length <= 0) && !this.isNoInsurance ) {
       invalid = true;
       error["noInsuranceTemplate"] = true;
+    }
+    else {
+      invalid = false;
+      error["noInsuranceTemplate"] = false;
     }
     switch (applicantType) {
       case "ResidentialTenant":
@@ -329,11 +338,26 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
           if (fileUpload.requiredDocumentType == Object.keys(this.RequiredDocumentTypes)[Object.values(this.RequiredDocumentTypes).indexOf(this.RequiredDocumentTypes.TenancyAgreement)])
             this.supportingDocumentsForm.get('hasCopyOfARentalAgreementOrLease').setValue(true);
           this.isLoading = false;
+          this._snackBar.open('Successfully uploaded the document', 'Close', {
+            horizontalPosition: 'center',
+            verticalPosition: 'top',
+            duration: 8000,
+          });
         },
         error: (error) => {
           console.error(error);
           this.isLoading = false;
-          document.location.href = 'https://dfa.gov.bc.ca/error.html';
+          //document.location.href = 'https://dfa.gov.bc.ca/error.html';
+          this._snackBar.open(
+              'Unable To upload the file. Please retry the upload process.',
+              'Close',
+              {
+                horizontalPosition: 'center',
+                verticalPosition: 'top',
+                duration: 8000,
+              }
+          );
+          
         }
       });
     } else {
@@ -358,11 +382,25 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
           fileUploads[requiredDocumentTypeFoundIndex] = fileUpload;
           this.formCreationService.fileUploadsForm.value.get('fileUploads').setValue(fileUploads);
           this.isLoading = false;
+          this._snackBar.open('Successfully uploaded the document', 'Close', {
+            horizontalPosition: 'center',
+            verticalPosition: 'top',
+            duration: 8000,
+          });
         },
         error: (error) => {
           console.error(error);
           this.isLoading = false;
-          document.location.href = 'https://dfa.gov.bc.ca/error.html';
+         // document.location.href = 'https://dfa.gov.bc.ca/error.html';
+          this._snackBar.open(
+            'Unable to upload the file. Please retry the upload process.',
+            'Close',
+            {
+              horizontalPosition: 'center',
+              verticalPosition: 'top',
+              duration: 8000,
+            }
+        );
         }
       });
     } else {
@@ -375,11 +413,25 @@ export default class SupportingDocumentsComponent implements OnInit, OnDestroy {
           if (fileUpload.requiredDocumentType == Object.keys(this.RequiredDocumentTypes)[Object.values(this.RequiredDocumentTypes).indexOf(this.RequiredDocumentTypes.TenancyAgreement)])
             this.supportingDocumentsForm.get('hasCopyOfARentalAgreementOrLease').setValue(true);
           this.isLoading = false;
+          this._snackBar.open('Successfully uploaded the document', 'Close', {
+            horizontalPosition: 'center',
+            verticalPosition: 'top',
+            duration: 8000,
+          });
         },
         error: (error) => {
           console.error(error);
           this.isLoading = false;
-          document.location.href = 'https://dfa.gov.bc.ca/error.html';
+          //document.location.href = 'https://dfa.gov.bc.ca/error.html';
+          this._snackBar.open(
+            'Unable to upload the file. Please retry the upload process.',
+            'Close',
+            {
+              horizontalPosition: 'center',
+              verticalPosition: 'top',
+              duration: 8000,
+            }
+        );
         }
       });
     }
