@@ -11,9 +11,9 @@ import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { FormCreationService } from '../../../../core/services/formCreation.service';
 import { Subscription } from 'rxjs';
 import { DirectivesModule } from '../../../../core/directives/directives.module';
-import { TextMaskModule } from 'angular2-text-mask';
 import * as globalConst from '../../../../core/services/globalConstants';
 import { Router } from '@angular/router';
+import { IMaskModule } from 'angular-imask';
 
 @Component({
   selector: 'app-personal-details',
@@ -88,7 +88,7 @@ export default class PersonalDetailsComponent implements OnInit, OnDestroy {
     MatInputModule,
     ReactiveFormsModule,
     DirectivesModule,
-    TextMaskModule,
+    IMaskModule,
     MatRadioModule
   ],
   declarations: [PersonalDetailsComponent]
