@@ -8,7 +8,8 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { DirectivesModule } from '../../../core/directives/directives.module';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { TextMaskModule } from 'angular2-text-mask';
+import { IMaskModule } from 'angular-imask';
+
 
 @NgModule({
   declarations: [PersonDetailFormComponent],
@@ -21,7 +22,7 @@ import { TextMaskModule } from 'angular2-text-mask';
     MatSelectModule,
     DirectivesModule,
     MatCheckboxModule,
-    TextMaskModule
+    IMaskModule
   ],
   exports: [PersonDetailFormComponent]
 })
